+++ src/components/common/Breadcrumb/titles.ts
@@ -1,4 +1,4 @@
-import { PATH_TO_TITLES } from '@/routes/Breadcrumb/pathToTitles';
+import { PATH_TO_TITLES } from '@/components/common/Breadcrumb/pathToTitles';
 import { PATH } from '@/routes/path';
 
 // titles 생성 (PATH_TO_TITLES의 경로에서 앞에 있는 / 제거)
